+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 v1.3.0
 ------
 
-* Adds of Alma badge feature on product pages
+* Adds an Alma badge with eligibility/payment plans information on product pages
 * Standardization of code
 
 v1.2.1